+++ bin/commands/on.rb
@@ -20,6 +20,7 @@ command :on do |c|
   add_options(:search, c)
   add_options(:tag_filter, c)
   add_options(:time_filter, c)
+  add_options(:save, c)
 
   c.action do |_global_options, options, args|
     if options[:output] && options[:output] !~ Doing::Plugins.plugin_regex(type: :export)
@@ -43,6 +44,12 @@ command :on do |c|
     options[:times] = true if options[:totals]
     options[:sort_tags] = options[:tag_sort]
 
-    Doing::Pager.page @wwid.list_date([start, finish], options[:section], options[:times], options[:output], options).chomp
+    Doing::Pager.page @wwid.list_date([start, finish],
+                                      options[:section],
+                                      options[:times],
+                                      options[:output],
+                                      options).chomp
+
+    Doing.config.save_view(options.to_view, options[:save].downcase) if options[:save]
   end
 end